+++ lib/src/job/task/parallel_executor.dart
@@ -1,0 +1,23 @@
+// Dart imports:
+import 'dart:async';
+
+// Package imports:
+import 'package:async_task/async_task.dart';
+
+class ParallelExecutor extends AsyncTask<void, void> {
+  @override
+  AsyncTask<void, void> instantiate(void parameters,
+      [Map<String, SharedData>? sharedData]) {
+    return this;
+  }
+
+  @override
+  void parameters() {}
+
+  @override
+  FutureOr<void> run() async {
+    try {} catch (e) {
+      rethrow;
+    }
+  }
+}